C99 でもビルド可能なように再構成
FreeBSD の mingw32-gcc (GCC) 4.8.1 でコンパイル可能か確認済み。

+++ src/mrbx_cstr_hexdigest.c
@@ -15,7 +15,7 @@ correct_bytesize(int bytesize)
 {
   if (bytesize < 1) {
     return sizeof(size_t);
-  } else if (bytesize > sizeof(uint64_t)) {
+  } else if ((size_t)bytesize > sizeof(uint64_t)) {
     return sizeof(uint64_t);
   } else {
     return bytesize;
